Change the way getTemplateString works
Before, I was allowing getTemplateString to be the overrideable
point for how to retrieve the string version of a template file.

Turns out, much easier to require the template directory at first
and still allow for the getTemplateString to be overridden as well.
To do that, we have to do a function check, and make sure to bind
that function call to the current 'this' value.